# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Desa Puundoho Information System.",
        "title": "Puundoho API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/penduduk/datasets": {
            "get": {
                "description": "Get a list of all population snapshots (years)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "List Population Datasets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.PendudukDataset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Initialize a new population year snapshot (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "Create Population Dataset",
                "parameters": [
                    {
                        "description": "Dataset Data",
                        "name": "dataset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PendudukDataset"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.PendudukDataset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/penduduk/datasets/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a whole year's population data (Admin only)",
                "tags": [
                    "penduduk"
                ],
                "summary": "Delete Population Dataset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/penduduk/datasets/{id}/bulk": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mass insert or update resident records for a dataset (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "Bulk Create Residents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of residents",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Penduduk"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/penduduk/datasets/{id}/records": {
            "get": {
                "description": "Get all resident records for a specific year snapshot",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "List Residents by Dataset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.Penduduk"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/penduduk/datasets/{id}/stats": {
            "get": {
                "description": "Get aggregated population data for charts and graphs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "Get Population Statistics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dataset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/penduduk/records/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a single resident record (Admin only)",
                "tags": [
                    "penduduk"
                ],
                "summary": "Delete Resident Record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update specific fields of a resident record (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "penduduk"
                ],
                "summary": "Patch Resident Record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Field Updates",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Penduduk": {
            "type": "object",
            "properties": {
                "agama": {
                    "type": "string"
                },
                "alamat": {
                    "type": "string"
                },
                "bisa_baca": {
                    "type": "boolean"
                },
                "dataset_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "kedudukan_keluarga": {
                    "type": "string"
                },
                "kewarganegaraan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "no_kk": {
                    "type": "string"
                },
                "pekerjaan": {
                    "type": "string"
                },
                "pend_terakhir": {
                    "type": "string"
                },
                "status_kawin": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tempat_lahir": {
                    "type": "string"
                }
            }
        },
        "main.PendudukDataset": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama_file": {
                    "type": "string"
                },
                "tahun": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Paste your token with \"Bearer \" prefix (e.g. \"Bearer \u003ctoken\u003e\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}